[LanguageRuntime] Move CPPLanguageRuntime into a plugin
Summary: This seems better suited to be in a plugin.

Reviewers: JDevlieghere, clayborg, jingham, compnerd, labath

Subscribers: mgorny, lldb-commits

Differential Revision: https://reviews.llvm.org/D64599

git-svn-id: https://llvm.org/svn/llvm-project/lldb/trunk@365951 91177308-0d34-0410-b5e6-96231b3b80d8
